pyproject.toml
[tool.poetry]
name = "python-fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = ["Takaaki Mizuno <takaaki.mizuno@gmail.com>"]
readme = "README.md"
packages = [{include = "python_fastapi_boilerplate"}]

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.10.3"
sqlalchemy = "^2.0.9"
python-dotenv = "^1.0.0"
fastapi = {extras = ["all"], version = "^0.95.0"}
injector = "^0.20.1"
click = "^8.1.3"
psycopg2 = "^2.9.6"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"
fastapi-auth-middleware = "^1.0.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
yapf = "^0.32.0"
toml = "^0.10.2"
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
